+++ tests/wl_tests_measures/test_measures_readability.py
@@ -45,22 +45,36 @@ test_text_eng_100 = Wl_Test_Text(TOKENS_MULTILEVEL_100)
 test_text_eng_120 = Wl_Test_Text(TOKENS_MULTILEVEL_120)
 test_text_eng_150 = Wl_Test_Text(TOKENS_MULTILEVEL_150)
 
+test_text_ara_0 = Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'ara')
+test_text_ara_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'ara')
+
+test_text_deu_0 = Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'deu_de')
+test_text_deu_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'deu_de')
+
+test_text_ita_0 = Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'ita')
+test_text_ita_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'ita')
+
 test_text_spa_0 = Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'spa')
 test_text_spa_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'spa')
 test_text_spa_120 = Wl_Test_Text(TOKENS_MULTILEVEL_120, lang = 'spa')
 test_text_spa_150 = Wl_Test_Text(TOKENS_MULTILEVEL_150, lang = 'spa')
 
+test_text_afr_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'afr')
+test_text_nld_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'nld')
+test_text_fra_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'fra')
+test_text_pol_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'pol')
+test_text_rus_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'rus')
 test_text_other_12 = Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'other')
 
 def test_automated_ara_readability_index():
-    aari_ara_0 = wl_measures_readability.automated_ara_readability_index(main, Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'ara'))
-    aari_ara_12 = wl_measures_readability.automated_ara_readability_index(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'ara'))
+    aari_ara_0 = wl_measures_readability.automated_ara_readability_index(main, test_text_ara_0)
+    aari_ara_12 = wl_measures_readability.automated_ara_readability_index(main, test_text_ara_12)
     aari_eng_12 = wl_measures_readability.automated_ara_readability_index(main, test_text_eng_12)
 
     print('Automated Arabic Readability Index:')
-    print(f'\t{aari_ara_0} (ara/0)')
-    print(f'\t{aari_ara_12} (ara/12)')
-    print(f'\t{aari_eng_12} (eng/12)')
+    print(f'\tara/0: {aari_ara_0}')
+    print(f'\tara/12: {aari_ara_12}')
+    print(f'\teng/12: {aari_eng_12}')
 
     assert aari_ara_0 == 'text_too_short'
     assert aari_ara_12 == 3.28 * 46 + 1.43 * (46 / 12) + 1.24 * (12 / 3)
@@ -72,9 +86,9 @@ def test_automated_readability_index():
     ari_spa_12 = wl_measures_readability.automated_readability_index(main, test_text_spa_12)
 
     print('Automated Readability Index:')
-    print(f'\t{ari_eng_0} (0)')
-    print(f'\t{ari_eng_12} (eng/12)')
-    print(f'\t{ari_spa_12} (spa/12)')
+    print(f'\teng/0: {ari_eng_0}')
+    print(f'\teng/12: {ari_eng_12}')
+    print(f'\tspa/12: {ari_spa_12}')
 
     assert ari_eng_0 == 'text_too_short'
     assert ari_eng_12 == ari_spa_12 == 0.5 * (12 / 3) + 4.71 * (47 / 12) - 21.43
@@ -85,9 +99,9 @@ def test_coleman_liau_index():
     grade_level_spa_12 = wl_measures_readability.coleman_liau_index(main, test_text_spa_12)
 
     print('Coleman-Liau Index:')
-    print(f'\t{grade_level_eng_0} (0)')
-    print(f'\t{grade_level_eng_12} (eng/12)')
-    print(f'\t{grade_level_spa_12} (spa/12)')
+    print(f'\teng/0: {grade_level_eng_0}')
+    print(f'\teng/12: {grade_level_eng_12}')
+    print(f'\tspa/12: {grade_level_spa_12}')
 
     est_cloze_pct = 141.8401 - 0.21459 * (45 / 12 * 100) + 1.079812 * (3 / 12 * 100)
 
@@ -100,9 +114,9 @@ def test_dale_chall_readability_score():
     x_c50_spa_12 = wl_measures_readability.dale_chall_readability_score(main, test_text_spa_12)
 
     print('Dale-Chall Readibility Score:')
-    print(f'\t{x_c50_eng_0} (0)')
-    print(f'\t{x_c50_eng_12} (eng/12)')
-    print(f'\t{x_c50_spa_12} (spa/12)')
+    print(f'\teng/0: {x_c50_eng_0}')
+    print(f'\teng/12: {x_c50_eng_12}')
+    print(f'\tspa/12: {x_c50_spa_12}')
 
     assert x_c50_eng_0 == 'text_too_short'
     assert x_c50_eng_12 == 0.1579 * (1 / 12) + 0.0496 * (12 / 3) + 3.6365
@@ -114,26 +128,12 @@ def test_devereux_readability_index():
     grade_placement_spa_12 = wl_measures_readability.devereux_readability_index(main, test_text_spa_12)
 
     print('Devereux Readability Index:')
-    print(f'\t{grade_placement_eng_0} (0)')
-    print(f'\t{grade_placement_eng_12} (eng/12)')
-    print(f'\t{grade_placement_spa_12} (spa/12)')
+    print(f'\teng/0: {grade_placement_eng_0}')
+    print(f'\teng/12: {grade_placement_eng_12}')
+    print(f'\tspa/12: {grade_placement_spa_12}')
 
     assert grade_placement_eng_0 == 'text_too_short'
     assert grade_placement_eng_12 == grade_placement_spa_12 == 1.56 * (47 / 12) + 0.19 * (12 / 3) - 6.49
-
-def test_fernandez_huertas_readability_score():
-    score_spa_0 = wl_measures_readability.fernandez_huertas_readability_score(main, test_text_spa_0)
-    score_spa_12 = wl_measures_readability.fernandez_huertas_readability_score(main, test_text_spa_12)
-    score_eng_12 = wl_measures_readability.fernandez_huertas_readability_score(main, test_text_eng_12)
-
-    print("Fernández Huerta's Readability Score:")
-    print(f'\t{score_spa_0} (spa/0)')
-    print(f'\t{score_spa_12} (spa/12)')
-    print(f'\t{score_eng_12} (eng/12)')
-
-    assert score_spa_0 == 'text_too_short'
-    assert score_spa_12 == 206.84 - 60 * (18 / 12) - 102 * (3 / 12)
-    assert score_eng_12 == 'no_support'
 
 def test_flesch_kincaid_grade_level():
     gl_eng_0 = wl_measures_readability.flesch_kincaid_grade_level(main, test_text_eng_0)
@@ -142,10 +142,10 @@ def test_flesch_kincaid_grade_level():
     gl_other_12 = wl_measures_readability.flesch_kincaid_grade_level(main, test_text_other_12)
 
     print('Flesch-Kincaid Grade Level:')
-    print(f'\t{gl_eng_0} (0)')
-    print(f'\t{gl_eng_12} (eng/12)')
-    print(f'\t{gl_spa_12} (spa/12)')
-    print(f'\t{gl_other_12} (other/12)')
+    print(f'\teng/0: {gl_eng_0}')
+    print(f'\teng/12: {gl_eng_12}')
+    print(f'\tspa/12: {gl_spa_12}')
+    print(f'\tother/12: {gl_other_12}')
 
     assert gl_eng_0 == 'text_too_short'
     assert gl_eng_12 == 0.39 * (12 / 3) + 11.8 * (15 / 12) - 15.59
@@ -155,18 +155,50 @@ def test_flesch_kincaid_grade_level():
 def test_flesch_reading_ease():
     flesch_re_eng_0 = wl_measures_readability.flesch_reading_ease(main, test_text_eng_0)
     flesch_re_eng_12 = wl_measures_readability.flesch_reading_ease(main, test_text_eng_12)
-    flesch_re_spa_12 = wl_measures_readability.flesch_reading_ease(main, test_text_spa_12)
+
+    main.settings_custom['measures']['readability']['re']['variant_nld'] = 'Douma'
+    flesch_re_nld_12_douma = wl_measures_readability.flesch_reading_ease(main, test_text_nld_12)
+    main.settings_custom['measures']['readability']['re']['variant_nld'] = "Brouwer's Leesindex A"
+    flesch_re_nld_12_brouwer = wl_measures_readability.flesch_reading_ease(main, test_text_nld_12)
+
+    flesch_re_fra_12 = wl_measures_readability.flesch_reading_ease(main, test_text_fra_12)
+    flesch_re_deu_12 = wl_measures_readability.flesch_reading_ease(main, test_text_deu_12)
+    flesch_re_ita_12 = wl_measures_readability.flesch_reading_ease(main, test_text_ita_12)
+    flesch_re_rus_12 = wl_measures_readability.flesch_reading_ease(main, test_text_rus_12)
+
+    main.settings_custom['measures']['readability']['re']['variant_spa'] = 'Fernández Huerta'
+    flesch_re_spa_12_fh = wl_measures_readability.flesch_reading_ease(main, test_text_spa_12)
+    main.settings_custom['measures']['readability']['re']['variant_spa'] = 'Szigriszt Pazos'
+    flesch_re_spa_12_sp = wl_measures_readability.flesch_reading_ease(main, test_text_spa_12)
+
+    flesch_re_afr_12 = wl_measures_readability.flesch_reading_ease(main, test_text_afr_12)
     flesch_re_other_12 = wl_measures_readability.flesch_reading_ease(main, test_text_other_12)
 
     print('Flesch Reading Ease:')
-    print(f'\t{flesch_re_eng_0} (0)')
-    print(f'\t{flesch_re_eng_12} (eng/12)')
-    print(f'\t{flesch_re_spa_12} (spa/12)')
-    print(f'\t{flesch_re_other_12} (other/12)')
+    print(f'\teng/0: {flesch_re_eng_0}')
+    print(f'\teng/12: {flesch_re_eng_12}')
+    print(f'\tnld-douma/12: {flesch_re_nld_12_douma}')
+    print(f'\tnld-brouwer/12: {flesch_re_nld_12_brouwer}')
+    print(f'\tfra/12: {flesch_re_fra_12}')
+    print(f'\tdeu/12: {flesch_re_deu_12}')
+    print(f'\tita/12: {flesch_re_ita_12}')
+    print(f'\trus/12: {flesch_re_rus_12}')
+    print(f'\tspa-fh/12: {flesch_re_spa_12_fh}')
+    print(f'\tspa-sp/12: {flesch_re_spa_12_sp}')
+    print(f'\tafr/12: {flesch_re_afr_12}')
+    print(f'\tother/12: {flesch_re_other_12}')
 
     assert flesch_re_eng_0 == 'text_too_short'
     assert flesch_re_eng_12 == 206.835 - 0.846 * (15 / 12 * 100) - 1.015 * (12 / 3)
-    assert flesch_re_spa_12 != 'no_support'
+    assert flesch_re_nld_12_douma == 206.84 - 77 * (18 / 12) - 0.93 * (12 / 3)
+    assert flesch_re_nld_12_brouwer == 195 - (200 / 3) * (18 / 12) - 2 * (12 / 3)
+    assert flesch_re_fra_12 == 207 - 73.6 * (16 / 12) - 1.015 * (12 / 3)
+    assert flesch_re_deu_12 == 180 - 58.5 * (15 / 12) - (12 / 3)
+    assert flesch_re_ita_12 == 217 - 60 * (19 / 12) - 1.3 * (12 / 3)
+    assert flesch_re_rus_12 == 206.835 - 60.1 * (13 / 12) - 1.3 * (12 / 3)
+    assert flesch_re_spa_12_fh == 206.84 - 60 * (18 / 12) - 1.02 * (12 / 3)
+    assert flesch_re_spa_12_sp == 206.84 - 62.3 * (18 / 12) - (12 / 3)
+    assert flesch_re_afr_12 == 206.835 - 0.846 * (18 / 12 * 100) - 1.015 * (12 / 3)
     assert flesch_re_other_12 == 'no_support'
 
 def test_flesch_reading_ease_simplified():
@@ -176,10 +208,10 @@ def test_flesch_reading_ease_simplified():
     flesch_re_simplified_other_12 = wl_measures_readability.flesch_reading_ease_simplified(main, test_text_other_12)
 
     print('Flesch Reading Ease (Simplified):')
-    print(f'\t{flesch_re_simplified_eng_0} (0)')
-    print(f'\t{flesch_re_simplified_eng_12} (eng/12)')
-    print(f'\t{flesch_re_simplified_spa_12} (spa/12)')
-    print(f'\t{flesch_re_simplified_other_12} (other/12)')
+    print(f'\teng/0: {flesch_re_simplified_eng_0}')
+    print(f'\teng/12: {flesch_re_simplified_eng_12}')
+    print(f'\tspa/12: {flesch_re_simplified_spa_12}')
+    print(f'\tother/12: {flesch_re_simplified_other_12}')
 
     assert flesch_re_simplified_eng_0 == 'text_too_short'
     assert flesch_re_simplified_eng_12 == flesch_re_simplified_spa_12 == 1.599 * (9 / 12 * 100) - 1.015 * (12 / 3) - 31.517
@@ -192,10 +224,10 @@ def test_forcast_grade_level():
     rgl_other_12 = wl_measures_readability.forcast_grade_level(main, test_text_other_12)
 
     print('FORCAST Grade Level:')
-    print(f'\t{rgl_eng_12} (eng/12)')
-    print(f'\t{rgl_eng_150} (eng/150)')
-    print(f'\t{rgl_spa_150} (spa/150)')
-    print(f'\t{rgl_other_12} (other/12)')
+    print(f'\teng/12: {rgl_eng_12}')
+    print(f'\teng/150: {rgl_eng_150}')
+    print(f'\tspa/150: {rgl_spa_150}')
+    print(f'\tother/12: {rgl_other_12}')
 
     assert rgl_eng_12 == 'text_too_short'
     assert rgl_eng_150 == rgl_spa_150 == 20.43 - 0.11 * (6 * 18 + 4)
@@ -207,9 +239,9 @@ def test_formula_de_comprensibilidad_de_gutierrez_de_polini():
     cp_eng_12 = wl_measures_readability.formula_de_comprensibilidad_de_gutierrez_de_polini(main, test_text_eng_12)
 
     print('Fórmula de comprensibilidad de Gutiérrez de Polini:')
-    print(f'\t{cp_spa_0} (spa/0)')
-    print(f'\t{cp_spa_12} (spa/12)')
-    print(f'\t{cp_eng_12} (eng/12)')
+    print(f'\tspa/0: {cp_spa_0}')
+    print(f'\tspa/12: {cp_spa_12}')
+    print(f'\teng/12: {cp_eng_12}')
 
     assert cp_spa_0 == 'text_too_short'
     assert cp_spa_12 == 95.2 - 9.7 * (45 / 12) - 0.35 * (12 / 3)
@@ -221,23 +253,23 @@ def test_formula_de_crawford():
     grade_level_eng_12 = wl_measures_readability.formula_de_crawford(main, test_text_eng_12)
 
     print('Fórmula de Crawford:')
-    print(f'\t{grade_level_spa_0} (spa/0)')
-    print(f'\t{grade_level_spa_12} (spa/12)')
-    print(f'\t{grade_level_eng_12} (eng/12)')
+    print(f'\tspa/0: {grade_level_spa_0}')
+    print(f'\tspa/12: {grade_level_spa_12}')
+    print(f'\teng/12: {grade_level_eng_12}')
 
     assert grade_level_spa_0 == 'text_too_short'
     assert grade_level_spa_12 == 3 / 12 * 100 * (-0.205) + 18 / 12 * 100 * 0.049 - 3.407
     assert grade_level_eng_12 == 'no_support'
 
 def test_gulpease_index():
-    gulpease_index_ita_0 = wl_measures_readability.gulpease_index(main, Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'ita'))
-    gulpease_index_ita_12 = wl_measures_readability.gulpease_index(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'ita'))
+    gulpease_index_ita_0 = wl_measures_readability.gulpease_index(main, test_text_ita_0)
+    gulpease_index_ita_12 = wl_measures_readability.gulpease_index(main, test_text_ita_12)
     gulpease_index_eng_12 = wl_measures_readability.gulpease_index(main, test_text_eng_12)
 
     print('Gulpease Index:')
-    print(f'\t{gulpease_index_ita_0} (ita/0)')
-    print(f'\t{gulpease_index_ita_12} (ita/12)')
-    print(f'\t{gulpease_index_eng_12} (eng/12)')
+    print(f'\tita/0: {gulpease_index_ita_0}')
+    print(f'\tita/12: {gulpease_index_ita_12}')
+    print(f'\teng/12: {gulpease_index_eng_12}')
 
     assert gulpease_index_ita_0 == 'text_too_short'
     assert gulpease_index_ita_12 == 89 + (300 * 3 - 10 * 45) / 12
@@ -246,14 +278,14 @@ def test_gulpease_index():
 def test_gunning_fog_index():
     fog_index_eng_0 = wl_measures_readability.gunning_fog_index(main, test_text_eng_0)
     fog_index_eng_12_propn = wl_measures_readability.gunning_fog_index(main, test_text_eng_12_propn)
-    fog_index_pol_12 = wl_measures_readability.gunning_fog_index(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'pol'))
+    fog_index_pol_12 = wl_measures_readability.gunning_fog_index(main, test_text_pol_12)
     fog_index_spa_12 = wl_measures_readability.gunning_fog_index(main, test_text_spa_12)
 
     print('Gunning Fog Index:')
-    print(f'\t{fog_index_eng_0} (0)')
-    print(f'\t{fog_index_eng_12_propn} (eng/12)')
-    print(f'\t{fog_index_pol_12} (pol/12)')
-    print(f'\t{fog_index_spa_12} (spa/12)')
+    print(f'\teng/0: {fog_index_eng_0}')
+    print(f'\teng/12: {fog_index_eng_12_propn}')
+    print(f'\tpol/12: {fog_index_pol_12}')
+    print(f'\tspa/12: {fog_index_spa_12}')
 
     assert fog_index_eng_0 == 'text_too_short'
     assert fog_index_eng_12_propn == 0.4 * (12 / 3 + 1 / 12 * 100)
@@ -266,9 +298,9 @@ def test_legibility_mu():
     mu_eng_12 = wl_measures_readability.legibility_mu(main, test_text_eng_12)
 
     print('Legibilidad µ:')
-    print(f'\t{mu_spa_0} (spa/0)')
-    print(f'\t{mu_spa_12} (spa/12)')
-    print(f'\t{mu_eng_12} (eng/12)')
+    print(f'\tspa/0: {mu_spa_0}')
+    print(f'\tspa/12: {mu_spa_12}')
+    print(f'\teng/12: {mu_eng_12}')
 
     assert mu_spa_0 == 'text_too_short'
     assert mu_spa_12 == (12 / 11) * (3.75 / 7.1875) * 100
@@ -281,10 +313,10 @@ def test_lensear_write():
     score_other_12 = wl_measures_readability.lensear_write(main, test_text_other_12)
 
     print('Lensear Write:')
-    print(f'\t{score_eng_0} (eng/0)')
-    print(f'\t{score_eng_12} (eng/12)')
-    print(f'\t{score_eng_100} (eng/100)')
-    print(f'\t{score_other_12} (other/12)')
+    print(f'\teng/0: {score_eng_0}')
+    print(f'\teng/12: {score_eng_12}')
+    print(f'\teng/100: {score_eng_100}')
+    print(f'\tother/12: {score_other_12}')
 
     assert score_eng_0 == 'text_too_short'
     assert score_eng_12 == 6 * (100 / 12) + 3 * 3 * (100 / 12)
@@ -297,9 +329,9 @@ def test_lix():
     lix_spa_12 = wl_measures_readability.lix(main, test_text_spa_12)
 
     print('Lix:')
-    print(f'\t{lix_eng_0} (eng/0)')
-    print(f'\t{lix_eng_12} (eng/12)')
-    print(f'\t{lix_spa_12} (spa/12)')
+    print(f'\teng/0: {lix_eng_0}')
+    print(f'\teng/12: {lix_eng_12}')
+    print(f'\tspa/12: {lix_spa_12}')
 
     assert lix_eng_0 == 'text_too_short'
     assert lix_eng_12 == 12 / 3 + 100 * (3 / 12)
@@ -311,23 +343,23 @@ def test_mcalpine_eflaw():
     eflaw_spa_12 = wl_measures_readability.mcalpine_eflaw(main, test_text_spa_12)
 
     print('McAlpine EFLAW Readability Score:')
-    print(f'\t{eflaw_eng_0} (eng/0)')
-    print(f'\t{eflaw_eng_12} (eng/12)')
-    print(f'\t{eflaw_spa_12} (spa/12)')
+    print(f'\teng/0: {eflaw_eng_0}')
+    print(f'\teng/12: {eflaw_eng_12}')
+    print(f'\tspa/12: {eflaw_spa_12}')
 
     assert eflaw_eng_0 == 'text_too_short'
     assert eflaw_eng_12 == (12 + 6) / 3
     assert eflaw_spa_12 == 'no_support'
 
 def test_osman():
-    osman_ara_0 = wl_measures_readability.osman(main, Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'ara'))
-    osman_ara_12 = wl_measures_readability.osman(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'ara'))
+    osman_ara_0 = wl_measures_readability.osman(main, test_text_ara_0)
+    osman_ara_12 = wl_measures_readability.osman(main, test_text_ara_12)
     osman_eng_12 = wl_measures_readability.osman(main, test_text_eng_12)
 
     print('OSMAN:')
-    print(f'\t{osman_ara_0} (ara/0)')
-    print(f'\t{osman_ara_12} (ara/12)')
-    print(f'\t{osman_eng_12} (eng/12)')
+    print(f'\tara/0: {osman_ara_0}')
+    print(f'\tara/12: {osman_ara_12}')
+    print(f'\teng/12: {osman_eng_12}')
 
     assert osman_ara_0 == 'text_too_short'
     assert osman_ara_12 == 200.791 - 1.015 * (12 / 3) - 24.181 * ((3 + 23 + 3 + 0) / 12)
@@ -339,9 +371,9 @@ def test_rix():
     rix_spa_12 = wl_measures_readability.rix(main, test_text_spa_12)
 
     print('Rix:')
-    print(f'\t{rix_eng_0} (eng/0)')
-    print(f'\t{rix_eng_12} (eng/12)')
-    print(f'\t{rix_spa_12} (spa/12)')
+    print(f'\teng/0: {rix_eng_0}')
+    print(f'\teng/12: {rix_eng_12}')
+    print(f'\tspa/12: {rix_spa_12}')
 
     assert rix_eng_0 == 'text_too_short'
     assert rix_eng_12 == rix_spa_12 == 3 / 3
@@ -353,10 +385,10 @@ def test_smog_grade():
     g_other_12 = wl_measures_readability.smog_grade(main, test_text_other_12)
 
     print('SMOG Grade:')
-    print(f'\t{g_eng_12} (eng/12)')
-    print(f'\t{g_eng_120} (eng/120)')
-    print(f'\t{g_spa_120} (spa/120)')
-    print(f'\t{g_other_12} (other/12)')
+    print(f'\teng/12: {g_eng_12}')
+    print(f'\teng/120: {g_eng_120}')
+    print(f'\tspa/120: {g_spa_120}')
+    print(f'\tother/12: {g_other_12}')
 
     assert g_eng_12 == 'text_too_short'
     assert g_eng_120 == 3.1291 + 1.043 * (15 ** 0.5)
@@ -369,43 +401,29 @@ def test_spache_grade_level():
     grade_level_spa_12 = wl_measures_readability.spache_grade_level(main, test_text_spa_12)
 
     print('Spache Grade Level:')
-    print(f'\t{grade_level_eng_12} (eng/12)')
-    print(f'\t{grade_level_eng_100} (eng/100)')
-    print(f'\t{grade_level_spa_12} (spa/12)')
+    print(f'\teng/12: {grade_level_eng_12}')
+    print(f'\teng/100: {grade_level_eng_100}')
+    print(f'\tspa/12: {grade_level_spa_12}')
 
     assert grade_level_eng_12 == 'text_too_short'
     assert grade_level_eng_100 == numpy.mean([0.141 * (100 / 25) + 0.086 * (25 / 100 * 100) + 0.839] * 3)
     assert grade_level_spa_12 == 'no_support'
 
-def test_szigriszts_perspicuity_index():
-    p_spa_0 = wl_measures_readability.szigriszts_perspicuity_index(main, test_text_spa_0)
-    p_spa_12 = wl_measures_readability.szigriszts_perspicuity_index(main, test_text_spa_12)
-    p_eng_12 = wl_measures_readability.szigriszts_perspicuity_index(main, test_text_eng_12)
-
-    print("Szigriszt's Perspicuity Index:")
-    print(f'\t{p_spa_0} (spa/0)')
-    print(f'\t{p_spa_12} (spa/12)')
-    print(f'\t{p_eng_12} (eng/12)')
-
-    assert p_spa_0 == 'text_too_short'
-    assert p_spa_12 == 207 - 62.3 * (18 / 12) - (12 / 3)
-    assert p_eng_12 == 'no_support'
-
 def test_wiener_sachtextformel():
-    wstf_deu_0 = wl_measures_readability.wiener_sachtextformel(main, Wl_Test_Text(TOKENS_MULTILEVEL_0, lang = 'deu_de'))
-    wstf_deu_12_1 = wl_measures_readability.wiener_sachtextformel(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'deu_de'), variant = '1')
-    wstf_deu_12_2 = wl_measures_readability.wiener_sachtextformel(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'deu_de'), variant = '2')
-    wstf_deu_12_3 = wl_measures_readability.wiener_sachtextformel(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'deu_de'), variant = '3')
-    wstf_deu_12_4 = wl_measures_readability.wiener_sachtextformel(main, Wl_Test_Text(TOKENS_MULTILEVEL_12, lang = 'deu_de'), variant = '4')
+    wstf_deu_0 = wl_measures_readability.wiener_sachtextformel(main, test_text_deu_0)
+    wstf_deu_12_1 = wl_measures_readability.wiener_sachtextformel(main, test_text_deu_12, variant = '1')
+    wstf_deu_12_2 = wl_measures_readability.wiener_sachtextformel(main, test_text_deu_12, variant = '2')
+    wstf_deu_12_3 = wl_measures_readability.wiener_sachtextformel(main, test_text_deu_12, variant = '3')
+    wstf_deu_12_4 = wl_measures_readability.wiener_sachtextformel(main, test_text_deu_12, variant = '4')
     wstf_eng_12 = wl_measures_readability.wiener_sachtextformel(main, test_text_eng_12)
 
     print('Wiener Sachtextformel:')
-    print(f'\t{wstf_deu_0} (deu/0)')
-    print(f'\t{wstf_deu_12_1} (deu-1/12)')
-    print(f'\t{wstf_deu_12_2} (deu-2/12)')
-    print(f'\t{wstf_deu_12_3} (deu-3/12)')
-    print(f'\t{wstf_deu_12_4} (deu-4/12)')
-    print(f'\t{wstf_eng_12} (eng/12)')
+    print(f'\tdeu/0: {wstf_deu_0}')
+    print(f'\tdeu-1/12: {wstf_deu_12_1}')
+    print(f'\tdeu-2/12: {wstf_deu_12_2}')
+    print(f'\tdeu-3/12: {wstf_deu_12_3}')
+    print(f'\tdeu-4/12: {wstf_deu_12_4}')
+    print(f'\teng/12: {wstf_eng_12}')
 
     ms = 0 / 12
     sl = 12 / 3
@@ -425,7 +443,6 @@ if __name__ == '__main__':
     test_coleman_liau_index()
     test_dale_chall_readability_score()
     test_devereux_readability_index()
-    test_fernandez_huertas_readability_score()
     test_flesch_kincaid_grade_level()
     test_flesch_reading_ease()
     test_flesch_reading_ease_simplified()
@@ -442,5 +459,4 @@ if __name__ == '__main__':
     test_rix()
     test_smog_grade()
     test_spache_grade_level()
-    test_szigriszts_perspicuity_index()
     test_wiener_sachtextformel()